requirements.txt
google-cloud-secret-manager==2.16.4
google-cloud-bigquery==3.11.4
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
pytz==2024.1
